fix user registration logic to ensure local file deletion only occurs if cover image path is provided; improve error handling for existing users

+++ src/controllers/user.controller.js
@@ -28,9 +28,11 @@ const registerUser = asynchandler(async (req, res) => {
     })
 
     if (existedUser) {
+
         // unlink uploaded file in local if user is trying to re-register
-        fs.unlinkSync(coverImageLocalPath)
+        if (coverImageLocalPath) fs.unlinkSync(coverImageLocalPath)
         fs.unlinkSync(avatarLocalPath)
+
         throw new ApiError(409, "User already exist")
     }
     //check for images, check of avatar
